fix: make sure client configs auth header gets refreshed
The access token that is returned by the user store is passed to the configuration used for the http client.

When the token needs to be refreshed we need to make sure we update the access token that is stored on the client instance making the request.

Otherwise for the life of that client it will just continue to use an expired access token.